server/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
redis
flask-cors
flask_marshmallow
flask_restful
Flask-JWT-Extended